requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
autopep8==2.3.2
click==8.3.1
dnspython==2.8.0
fastapi==0.135.1
h11==0.16.0
httptools==0.7.1
idna==3.11
motor==3.7.1
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
groq
